Quickfix for Series 13 moves added, more adjustments
TL;DR: Series 13 relevant mythical moves are added, switching to Porydex for Usage Stats link, and fixing the outdated names of some Pokemon forms which means all Trainer Tower sets with them will no longer appear.

In-depth Details:

The following moves have been added for use in the Series 13 metagame calculations:
-Techno Blast
-Steam Eruption
-Double Iron Bash
-Jungle Healing
Series 13 sets will be added later.

The Usages Stats link now directs to Porydex. This is due to the nature of the damage calculator's legacy support, which Porydex provides a more in depth look, going month by month in usage, than Pikalytics, which only shows the final usage stats of each format.

A number of Pokemon forms have outdated short-hand terms that could lead to confusion when using the damage calculator, so they're in the process of being updated. This will affect the sets that Trainer Tower had, and as such they will not appear any longer for these forms. This is part of the process of the damage calculator undergoing some reworkings to make future generational updates easier and quicker, as well as updating some old sets still present from the Trainer Tower days.