Scarlet and Violet support
The Damage Calculator is now supporting Pokemon Scarlet and Violet! Everything from new Pokemon, items, abilities, and moves have been added to the calculator. Expect at least one update in the following week or so, as we're still missing information on some things, which can lead to inaccurate calculations.

To celebrate the game's release, the panels for gen 9 have been changed to match scarlet and violet colors! The current plan is for these to stay until the end of the month, but that can change to be sooner/later depending on reception.

There are currently no sets in the calc, as I do not want to take any guesses before even touching the games.