IV restrictions on Pokemon with fixed IVs
Added restrictions on IVs for Pokemon that only have fixed IVs currently. Affected Pokemon:
- Koraidon
- Miraidon
- Ogerpon
- Gouging Fire
- Raging Bolt
- Iron Boulder
- Iron Crown
- Terapagos

This only affects the minimum for each of the IVs and the numbers in between the minimum and 31. Can currently be bypassed by typing the IV field instead. I might revert if it ends up being more annoying than useful or if it causes other problems not caught from testing.