Custom modifiers for mechanics testing
Added a togglable option for custom modifiers. Supports application to Base Power, Attack, Defense, and Final modifiers. Input is in terms of [value]/4096 for complete accuracy, since that is how it's handled in-game. While the normal text output doesn't list the modifiers, they are listed below the number roll values.

This is intended mostly for mechanics testing for a new game's release, or a hypothetical balance patch should that ever happen. It's only active if the user explicitly turns it on, otherwise it doesn't display.